feat: guard deletion from arbitrary users

+++ server/src/item/item.controller.ts
@@ -6,13 +6,20 @@ import {
   Get,
   Param,
   UseGuards,
+  UnauthorizedException,
+  NotFoundException,
+  Logger,
 } from '@nestjs/common';
 import { ItemService } from './item.service';
 import { CreateItemDto } from './dto/createItem.dto';
 import { JwtAuthGuard } from '../auth/guards/jwt.guard';
+import { User } from 'src/common/decorators/user.decorator';
+import type { UserPayload } from 'src/common/interfaces/user_payload.interface';
 
 @Controller('items')
 export class ItemController {
+  private readonly logger = new Logger(ItemController.name);
+
   constructor(private readonly itemService: ItemService) {}
 
   @Post()
@@ -65,7 +72,19 @@ export class ItemController {
 
   @Delete(':id')
   @UseGuards(JwtAuthGuard)
-  async deleteItem(@Param('id') id: string): Promise<any> {
+  async deleteItem(
+    @Param('id') id: string,
+    @User() user: UserPayload,
+  ): Promise<any> {
+    const itemToRemove = await this.itemService.findItemById(parseInt(id, 10));
+    if (!itemToRemove) {
+      throw new NotFoundException('Item not found.');
+    }
+    if (user.userId != itemToRemove?.user?.id && user.role != 'admin') {
+      throw new UnauthorizedException(
+        'User is unauthorized to do this action.',
+      );
+    }
     await this.itemService.deleteItem(parseInt(id, 10));
     return {
       message: 'Item deleted successfully',
